[TASK] Port to Thin Client
Add compile time flag for tty.rs to assume a device under /dev/ttyACM0
is a Bluegiga Dongle if running on a 32-bit Linux.
At this time, Thin Client udev library does not expose enough parameters
for tty.rs to recognize Bluegiga Dongle.
The purpose of this change is to allow creating a testable build running
on Thin Client.